Add jvmArgs and log to gradle tests output.

+++ app/src/test/resources/src/TestSumLoop.java
@@ -2,8 +2,8 @@ public class TestSumLoop {
 
   private static int[] arr = {1, 2, 3, 4, 5, 6};
 
-  public static double compute() {
-    double sum = 0;
+  public static int compute() {
+    int sum = 0;
     for (int i = 0; i < arr.length; i++) {
       sum += arr[i];
     }
